backend_py/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
aiomysql==0.2.0
python-dotenv==1.0.1
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
